Move deploy commands to be near commands, add index exports
deploy command init with all slash commands ans types for init 4
