+++ src/commands/DeployCommands.ts
@@ -1,26 +1,25 @@
-import config from './config'
+import config from '../utils/dev/config'
 import { REST } from '@discordjs/rest'
 import { Routes } from 'discord-api-types/v9'
-import { sendAlertToConsole } from './sendAlertToConsole'
+import { sendAlertToConsole } from '../utils/_index'
+import { AllSlashCommands } from './_index'
 
-import type { Command } from '../../interfaces/Command'
+import type { CombinedCommands } from '../interfaces/Command'
 
 const {token, client, guild} = config;
 
 const rest = new REST({ version: '9' }).setToken(token!);
 
-export const DeployCommands = async (): Promise<Command[]> => {
+export const DeployCommands = async (): Promise<CombinedCommands[]> => {
   try {
     sendAlertToConsole(`Updating Quebert's application (/) commands. Brb.`)
     await rest.put(
       Routes.applicationGuildCommands(client!, guild!),
-      { body: ['see below']}
-      //! { body: AllSlashCommands.map(c => c.data.toJSON()) },
+      { body: AllSlashCommands.map(c => c.data.toJSON()) }
     )
     sendAlertToConsole(`Im back and successfully updated Quebert's application (/) commands!`)
-    return ([]) //! should be an array of commands, not optional
   } catch (error) {
-    console.error(error)
+    sendAlertToConsole(error)
   }
-  return([]) //! Same here as above
+  return(AllSlashCommands)
 }
